Module aws_sdk_macie2::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptInvitation

Fluent builder constructing a request to AcceptInvitation.

BatchGetCustomDataIdentifiers

Fluent builder constructing a request to BatchGetCustomDataIdentifiers.

CreateClassificationJob

Fluent builder constructing a request to CreateClassificationJob.

CreateCustomDataIdentifier

Fluent builder constructing a request to CreateCustomDataIdentifier.

CreateFindingsFilter

Fluent builder constructing a request to CreateFindingsFilter.

CreateInvitations

Fluent builder constructing a request to CreateInvitations.

CreateMember

Fluent builder constructing a request to CreateMember.

CreateSampleFindings

Fluent builder constructing a request to CreateSampleFindings.

DeclineInvitations

Fluent builder constructing a request to DeclineInvitations.

DeleteCustomDataIdentifier

Fluent builder constructing a request to DeleteCustomDataIdentifier.

DeleteFindingsFilter

Fluent builder constructing a request to DeleteFindingsFilter.

DeleteInvitations

Fluent builder constructing a request to DeleteInvitations.

DeleteMember

Fluent builder constructing a request to DeleteMember.

DescribeBuckets

Fluent builder constructing a request to DescribeBuckets.

DescribeClassificationJob

Fluent builder constructing a request to DescribeClassificationJob.

DescribeOrganizationConfiguration

Fluent builder constructing a request to DescribeOrganizationConfiguration.

DisableMacie

Fluent builder constructing a request to DisableMacie.

DisableOrganizationAdminAccount

Fluent builder constructing a request to DisableOrganizationAdminAccount.

DisassociateFromAdministratorAccount

Fluent builder constructing a request to DisassociateFromAdministratorAccount.

DisassociateFromMasterAccount

Fluent builder constructing a request to DisassociateFromMasterAccount.

DisassociateMember

Fluent builder constructing a request to DisassociateMember.

EnableMacie

Fluent builder constructing a request to EnableMacie.

EnableOrganizationAdminAccount

Fluent builder constructing a request to EnableOrganizationAdminAccount.

GetAdministratorAccount

Fluent builder constructing a request to GetAdministratorAccount.

GetBucketStatistics

Fluent builder constructing a request to GetBucketStatistics.

GetClassificationExportConfiguration

Fluent builder constructing a request to GetClassificationExportConfiguration.

GetCustomDataIdentifier

Fluent builder constructing a request to GetCustomDataIdentifier.

GetFindingStatistics

Fluent builder constructing a request to GetFindingStatistics.

GetFindings

Fluent builder constructing a request to GetFindings.

GetFindingsFilter

Fluent builder constructing a request to GetFindingsFilter.

GetFindingsPublicationConfiguration

Fluent builder constructing a request to GetFindingsPublicationConfiguration.

GetInvitationsCount

Fluent builder constructing a request to GetInvitationsCount.

GetMacieSession

Fluent builder constructing a request to GetMacieSession.

GetMasterAccount

Fluent builder constructing a request to GetMasterAccount.

GetMember

Fluent builder constructing a request to GetMember.

GetUsageStatistics

Fluent builder constructing a request to GetUsageStatistics.

GetUsageTotals

Fluent builder constructing a request to GetUsageTotals.

ListClassificationJobs

Fluent builder constructing a request to ListClassificationJobs.

ListCustomDataIdentifiers

Fluent builder constructing a request to ListCustomDataIdentifiers.

ListFindings

Fluent builder constructing a request to ListFindings.

ListFindingsFilters

Fluent builder constructing a request to ListFindingsFilters.

ListInvitations

Fluent builder constructing a request to ListInvitations.

ListManagedDataIdentifiers

Fluent builder constructing a request to ListManagedDataIdentifiers.

ListMembers

Fluent builder constructing a request to ListMembers.

ListOrganizationAdminAccounts

Fluent builder constructing a request to ListOrganizationAdminAccounts.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

PutClassificationExportConfiguration

Fluent builder constructing a request to PutClassificationExportConfiguration.

PutFindingsPublicationConfiguration

Fluent builder constructing a request to PutFindingsPublicationConfiguration.

SearchResources

Fluent builder constructing a request to SearchResources.

TagResource

Fluent builder constructing a request to TagResource.

TestCustomDataIdentifier

Fluent builder constructing a request to TestCustomDataIdentifier.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateClassificationJob

Fluent builder constructing a request to UpdateClassificationJob.

UpdateFindingsFilter

Fluent builder constructing a request to UpdateFindingsFilter.

UpdateMacieSession

Fluent builder constructing a request to UpdateMacieSession.

UpdateMemberSession

Fluent builder constructing a request to UpdateMemberSession.

UpdateOrganizationConfiguration

Fluent builder constructing a request to UpdateOrganizationConfiguration.