Struct aws_sdk_macie2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Macie2.

This client allows ergonomic access to a Macie2-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_invitation(&self) -> AcceptInvitation<C, M, R>

pub fn batch_get_custom_data_identifiers(
&self
) -> BatchGetCustomDataIdentifiers<C, M, R>

pub fn create_classification_job(&self) -> CreateClassificationJob<C, M, R>

pub fn create_custom_data_identifier(
&self
) -> CreateCustomDataIdentifier<C, M, R>

pub fn create_findings_filter(&self) -> CreateFindingsFilter<C, M, R>

pub fn create_invitations(&self) -> CreateInvitations<C, M, R>

pub fn create_member(&self) -> CreateMember<C, M, R>

pub fn create_sample_findings(&self) -> CreateSampleFindings<C, M, R>

pub fn decline_invitations(&self) -> DeclineInvitations<C, M, R>

pub fn delete_custom_data_identifier(
&self
) -> DeleteCustomDataIdentifier<C, M, R>

pub fn delete_findings_filter(&self) -> DeleteFindingsFilter<C, M, R>

pub fn delete_invitations(&self) -> DeleteInvitations<C, M, R>

pub fn delete_member(&self) -> DeleteMember<C, M, R>

pub fn describe_buckets(&self) -> DescribeBuckets<C, M, R>

pub fn describe_classification_job(&self) -> DescribeClassificationJob<C, M, R>

pub fn describe_organization_configuration(
&self
) -> DescribeOrganizationConfiguration<C, M, R>

pub fn disable_macie(&self) -> DisableMacie<C, M, R>

pub fn disable_organization_admin_account(
&self
) -> DisableOrganizationAdminAccount<C, M, R>

pub fn disassociate_from_administrator_account(
&self
) -> DisassociateFromAdministratorAccount<C, M, R>

pub fn disassociate_from_master_account(
&self
) -> DisassociateFromMasterAccount<C, M, R>

pub fn disassociate_member(&self) -> DisassociateMember<C, M, R>

pub fn enable_macie(&self) -> EnableMacie<C, M, R>

pub fn enable_organization_admin_account(
&self
) -> EnableOrganizationAdminAccount<C, M, R>

pub fn get_administrator_account(&self) -> GetAdministratorAccount<C, M, R>

pub fn get_bucket_statistics(&self) -> GetBucketStatistics<C, M, R>

pub fn get_classification_export_configuration(
&self
) -> GetClassificationExportConfiguration<C, M, R>

pub fn get_custom_data_identifier(&self) -> GetCustomDataIdentifier<C, M, R>

pub fn get_findings(&self) -> GetFindings<C, M, R>

pub fn get_findings_filter(&self) -> GetFindingsFilter<C, M, R>

pub fn get_findings_publication_configuration(
&self
) -> GetFindingsPublicationConfiguration<C, M, R>

pub fn get_finding_statistics(&self) -> GetFindingStatistics<C, M, R>

pub fn get_invitations_count(&self) -> GetInvitationsCount<C, M, R>

pub fn get_macie_session(&self) -> GetMacieSession<C, M, R>

pub fn get_master_account(&self) -> GetMasterAccount<C, M, R>

pub fn get_member(&self) -> GetMember<C, M, R>

pub fn get_usage_statistics(&self) -> GetUsageStatistics<C, M, R>

pub fn get_usage_totals(&self) -> GetUsageTotals<C, M, R>

pub fn list_classification_jobs(&self) -> ListClassificationJobs<C, M, R>

pub fn list_custom_data_identifiers(&self) -> ListCustomDataIdentifiers<C, M, R>

pub fn list_findings(&self) -> ListFindings<C, M, R>

pub fn list_findings_filters(&self) -> ListFindingsFilters<C, M, R>

pub fn list_invitations(&self) -> ListInvitations<C, M, R>

pub fn list_members(&self) -> ListMembers<C, M, R>

pub fn list_organization_admin_accounts(
&self
) -> ListOrganizationAdminAccounts<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn put_classification_export_configuration(
&self
) -> PutClassificationExportConfiguration<C, M, R>

pub fn put_findings_publication_configuration(
&self
) -> PutFindingsPublicationConfiguration<C, M, R>

pub fn search_resources(&self) -> SearchResources<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_custom_data_identifier(&self) -> TestCustomDataIdentifier<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_classification_job(&self) -> UpdateClassificationJob<C, M, R>

pub fn update_findings_filter(&self) -> UpdateFindingsFilter<C, M, R>

pub fn update_macie_session(&self) -> UpdateMacieSession<C, M, R>

pub fn update_member_session(&self) -> UpdateMemberSession<C, M, R>

pub fn update_organization_configuration(
&self
) -> UpdateOrganizationConfiguration<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self