Module aws_sdk_macie2::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AcceptInvitationError

AccessDeniedException

Provides information about an error that occurred due to insufficient access to a specified resource.

BatchGetCustomDataIdentifiersError

ConflictException

Provides information about an error that occurred due to a versioning conflict for a specified resource.

CreateClassificationJobError

CreateCustomDataIdentifierError

CreateFindingsFilterError

CreateInvitationsError

CreateMemberError

CreateSampleFindingsError

DeclineInvitationsError

DeleteCustomDataIdentifierError

DeleteFindingsFilterError

DeleteInvitationsError

DeleteMemberError

DescribeBucketsError

DescribeClassificationJobError

DescribeOrganizationConfigurationError

DisableMacieError

DisableOrganizationAdminAccountError

DisassociateFromAdministratorAccountError

DisassociateFromMasterAccountError

DisassociateMemberError

EnableMacieError

EnableOrganizationAdminAccountError

GetAdministratorAccountError

GetBucketStatisticsError

GetClassificationExportConfigurationError

GetCustomDataIdentifierError

GetFindingStatisticsError

GetFindingsError

GetFindingsFilterError

GetFindingsPublicationConfigurationError

GetInvitationsCountError

GetMacieSessionError

GetMasterAccountError

GetMemberError

GetUsageStatisticsError

GetUsageTotalsError

InternalServerException

Provides information about an error that occurred due to an unknown internal server error, exception, or failure.

ListClassificationJobsError

ListCustomDataIdentifiersError

ListFindingsError

ListFindingsFiltersError

ListInvitationsError

ListMembersError

ListOrganizationAdminAccountsError

ListTagsForResourceError

PutClassificationExportConfigurationError

PutFindingsPublicationConfigurationError

ResourceNotFoundException

Provides information about an error that occurred because a specified resource wasn't found.

SearchResourcesError

ServiceQuotaExceededException

Provides information about an error that occurred due to one or more service quotas for an account.

TagResourceError

TestCustomDataIdentifierError

ThrottlingException

Provides information about an error that occurred because too many requests were sent during a certain amount of time.

UntagResourceError

UpdateClassificationJobError

UpdateFindingsFilterError

UpdateMacieSessionError

UpdateMemberSessionError

UpdateOrganizationConfigurationError

ValidationException

Provides information about an error that occurred due to a syntax error in a request.

Enums

AcceptInvitationErrorKind

BatchGetCustomDataIdentifiersErrorKind

CreateClassificationJobErrorKind

CreateCustomDataIdentifierErrorKind

CreateFindingsFilterErrorKind

CreateInvitationsErrorKind

CreateMemberErrorKind

CreateSampleFindingsErrorKind

DeclineInvitationsErrorKind

DeleteCustomDataIdentifierErrorKind

DeleteFindingsFilterErrorKind

DeleteInvitationsErrorKind

DeleteMemberErrorKind

DescribeBucketsErrorKind

DescribeClassificationJobErrorKind

DescribeOrganizationConfigurationErrorKind

DisableMacieErrorKind

DisableOrganizationAdminAccountErrorKind

DisassociateFromAdministratorAccountErrorKind

DisassociateFromMasterAccountErrorKind

DisassociateMemberErrorKind

EnableMacieErrorKind

EnableOrganizationAdminAccountErrorKind

GetAdministratorAccountErrorKind

GetBucketStatisticsErrorKind

GetClassificationExportConfigurationErrorKind

GetCustomDataIdentifierErrorKind

GetFindingStatisticsErrorKind

GetFindingsErrorKind

GetFindingsFilterErrorKind

GetFindingsPublicationConfigurationErrorKind

GetInvitationsCountErrorKind

GetMacieSessionErrorKind

GetMasterAccountErrorKind

GetMemberErrorKind

GetUsageStatisticsErrorKind

GetUsageTotalsErrorKind

ListClassificationJobsErrorKind

ListCustomDataIdentifiersErrorKind

ListFindingsErrorKind

ListFindingsFiltersErrorKind

ListInvitationsErrorKind

ListMembersErrorKind

ListOrganizationAdminAccountsErrorKind

ListTagsForResourceErrorKind

PutClassificationExportConfigurationErrorKind

PutFindingsPublicationConfigurationErrorKind

SearchResourcesErrorKind

TagResourceErrorKind

TestCustomDataIdentifierErrorKind

UntagResourceErrorKind

UpdateClassificationJobErrorKind

UpdateFindingsFilterErrorKind

UpdateMacieSessionErrorKind

UpdateMemberSessionErrorKind

UpdateOrganizationConfigurationErrorKind