Struct aws_sdk_macie2::input::create_classification_job_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateClassificationJobInput

Implementations

impl Builder

pub fn client_token(self, input: impl Into<String>) -> Self

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn custom_data_identifier_ids(self, input: impl Into<String>) -> Self

pub fn set_custom_data_identifier_ids(self, input: Option<Vec<String>>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A custom description of the job. The description can contain as many as 200 characters.

pub fn set_description(self, input: Option<String>) -> Self

pub fn initial_run(self, input: bool) -> Self

Specifies whether to analyze all existing, eligible objects immediately after the job is created.

pub fn set_initial_run(self, input: Option<bool>) -> Self

pub fn job_type(self, input: JobType) -> Self

The schedule for running the job. Valid values are:

ONE_TIME - Run the job only once. If you specify this value, don't specify a value for the scheduleFrequency property.

SCHEDULED - Run the job on a daily, weekly, or monthly basis. If you specify this value, use the scheduleFrequency property to define the recurrence pattern for the job.

pub fn set_job_type(self, input: Option<JobType>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A custom name for the job. The name can contain as many as 500 characters.

pub fn set_name(self, input: Option<String>) -> Self

pub fn s3_job_definition(self, input: S3JobDefinition) -> Self

The S3 buckets that contain the objects to analyze, and the scope of that analysis.

pub fn set_s3_job_definition(self, input: Option<S3JobDefinition>) -> Self

pub fn sampling_percentage(self, input: i32) -> Self

The sampling depth, as a percentage, to apply when processing objects. This value determines the percentage of eligible objects that the job analyzes. If this value is less than 100, Amazon Macie selects the objects to analyze at random, up to the specified percentage, and analyzes all the data in those objects.

pub fn set_sampling_percentage(self, input: Option<i32>) -> Self

pub fn schedule_frequency(self, input: JobScheduleFrequency) -> Self

The recurrence pattern for running the job. To run the job only once, don't specify a value for this property and set the value for the jobType property to ONE_TIME.

pub fn set_schedule_frequency(self, input: Option<JobScheduleFrequency>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn build(self) -> Result<CreateClassificationJobInput, BuildError>

Consumes the builder and constructs a CreateClassificationJobInput