Module aws_sdk_macie2::model

Modules

access_control_list

See AccessControlList

account_detail

See AccountDetail

account_level_permissions

See AccountLevelPermissions

admin_account

See AdminAccount

api_call_details

See ApiCallDetails

assumed_role

See AssumedRole

aws_account

See AwsAccount

aws_service

See AwsService

batch_get_custom_data_identifier_summary

See BatchGetCustomDataIdentifierSummary

block_public_access

See BlockPublicAccess

bucket_count_by_effective_permission

See BucketCountByEffectivePermission

bucket_count_by_encryption_type

See BucketCountByEncryptionType

bucket_count_by_shared_access_type

See BucketCountBySharedAccessType

bucket_count_policy_allows_unencrypted_object_uploads

See BucketCountPolicyAllowsUnencryptedObjectUploads

bucket_criteria_additional_properties

See BucketCriteriaAdditionalProperties

bucket_level_permissions

See BucketLevelPermissions

bucket_metadata

See BucketMetadata

bucket_permission_configuration

See BucketPermissionConfiguration

bucket_policy

See BucketPolicy

bucket_public_access

See BucketPublicAccess

bucket_server_side_encryption

See BucketServerSideEncryption

bucket_sort_criteria

See BucketSortCriteria

cell

See Cell

classification_details

See ClassificationDetails

classification_export_configuration

See ClassificationExportConfiguration

classification_result

See ClassificationResult

classification_result_status

See ClassificationResultStatus

criteria_block_for_job

See CriteriaBlockForJob

criteria_for_job

See CriteriaForJob

criterion_additional_properties

See CriterionAdditionalProperties

custom_data_identifier_summary

See CustomDataIdentifierSummary

custom_data_identifiers

See CustomDataIdentifiers

custom_detection

See CustomDetection

daily_schedule

See DailySchedule

default_detection

See DefaultDetection

domain_details

See DomainDetails

federated_user

See FederatedUser

finding

See Finding

finding_action

See FindingAction

finding_actor

See FindingActor

finding_criteria

See FindingCriteria

finding_statistics_sort_criteria

See FindingStatisticsSortCriteria

findings_filter_list_item

See FindingsFilterListItem

group_count

See GroupCount

iam_user

See IamUser

invitation

See Invitation

ip_address_details

See IpAddressDetails

ip_city

See IpCity

ip_country

See IpCountry

ip_geo_location

See IpGeoLocation

ip_owner

See IpOwner

job_details

See JobDetails

job_schedule_frequency

See JobScheduleFrequency

job_scope_term

See JobScopeTerm

job_scoping_block

See JobScopingBlock

job_summary

See JobSummary

key_value_pair

See KeyValuePair

last_run_error_status

See LastRunErrorStatus

list_jobs_filter_criteria

See ListJobsFilterCriteria

list_jobs_filter_term

See ListJobsFilterTerm

list_jobs_sort_criteria

See ListJobsSortCriteria

matching_bucket

See MatchingBucket

matching_resource

See MatchingResource

member

See Member

monthly_schedule

See MonthlySchedule

object_count_by_encryption_type

See ObjectCountByEncryptionType

object_level_statistics

See ObjectLevelStatistics

occurrences

See Occurrences

page

See Page

policy_details

See PolicyDetails

range

See Range

record

See Record

replication_details

See ReplicationDetails

resources_affected

See ResourcesAffected

s3_bucket

See S3Bucket

s3_bucket_criteria_for_job

See S3BucketCriteriaForJob

s3_bucket_definition_for_job

See S3BucketDefinitionForJob

s3_bucket_owner

See S3BucketOwner

s3_destination

See S3Destination

s3_job_definition

See S3JobDefinition

s3_object

See S3Object

scoping

See Scoping

search_resources_bucket_criteria

See SearchResourcesBucketCriteria

search_resources_criteria

See SearchResourcesCriteria

search_resources_criteria_block

See SearchResourcesCriteriaBlock

search_resources_simple_criterion

See SearchResourcesSimpleCriterion

search_resources_sort_criteria

See SearchResourcesSortCriteria

search_resources_tag_criterion

See SearchResourcesTagCriterion

search_resources_tag_criterion_pair

See SearchResourcesTagCriterionPair

security_hub_configuration

See SecurityHubConfiguration

sensitive_data_item

See SensitiveDataItem

server_side_encryption

See ServerSideEncryption

service_limit

See ServiceLimit

session_context

See SessionContext

session_context_attributes

See SessionContextAttributes

session_issuer

See SessionIssuer

severity

See Severity

simple_criterion_for_job

See SimpleCriterionForJob

simple_scope_term

See SimpleScopeTerm

sort_criteria

See SortCriteria

statistics

See Statistics

tag_criterion_for_job

See TagCriterionForJob

tag_criterion_pair_for_job

See TagCriterionPairForJob

tag_scope_term

See TagScopeTerm

tag_value_pair

See TagValuePair

unprocessed_account

See UnprocessedAccount

usage_by_account

See UsageByAccount

usage_record

See UsageRecord

usage_statistics_filter

See UsageStatisticsFilter

usage_statistics_sort_by

See UsageStatisticsSortBy

usage_total

See UsageTotal

user_identity

See UserIdentity

user_identity_root

See UserIdentityRoot

user_paused_details

See UserPausedDetails

weekly_schedule

See WeeklySchedule

Structs

AccessControlList

Provides information about the permissions settings of the bucket-level access control list (ACL) for an S3 bucket.

AccountDetail

Specifies the details of an account to associate with an Amazon Macie administrator account.

AccountLevelPermissions

Provides information about the account-level permissions settings that apply to an S3 bucket.

AdminAccount

Provides information about the delegated Amazon Macie administrator account for an Amazon Web Services organization.

ApiCallDetails

Provides information about an API operation that an entity invoked for an affected resource.

AssumedRole

Provides information about an identity that performed an action on an affected resource by using temporary security credentials. The credentials were obtained using the AssumeRole operation of the Security Token Service (STS) API.

AwsAccount

Provides information about an Amazon Web Services account and entity that performed an action on an affected resource. The action was performed using the credentials for an Amazon Web Services account other than your own account.

AwsService

Provides information about an Amazon Web Service that performed an action on an affected resource.

BatchGetCustomDataIdentifierSummary

Provides information about a custom data identifier.

BlockPublicAccess

Provides information about the block public access settings for an S3 bucket. These settings can apply to a bucket at the account level or bucket level. For detailed information about each setting, see Blocking public access to your Amazon S3 storage in the Amazon Simple Storage Service User Guide.

BucketCountByEffectivePermission

Provides information about the number of S3 buckets that are publicly accessible based on a combination of permissions settings for each bucket.

BucketCountByEncryptionType

Provides information about the number of S3 buckets that use certain types of server-side encryption by default or don't encrypt new objects by default. For detailed information about these settings, see Setting default server-side encryption behavior for Amazon S3 buckets in the Amazon Simple Storage Service User Guide.

BucketCountBySharedAccessType

Provides information about the number of S3 buckets that are or aren't shared with other Amazon Web Services accounts.

BucketCountPolicyAllowsUnencryptedObjectUploads

Provides information about the number of S3 buckets whose bucket policies do or don't require server-side encryption of objects when objects are uploaded to the buckets.

BucketCriteriaAdditionalProperties

Specifies the operator to use in a property-based condition that filters the results of a query for information about S3 buckets.

BucketLevelPermissions

Provides information about the bucket-level permissions settings for an S3 bucket.

BucketMetadata

Provides information about an S3 bucket that Amazon Macie monitors and analyzes.

BucketPermissionConfiguration

Provides information about the account-level and bucket-level permissions settings for an S3 bucket.

BucketPolicy

Provides information about the permissions settings of the bucket policy for an S3 bucket.

BucketPublicAccess

Provides information about the permissions settings that determine whether an S3 bucket is publicly accessible.

BucketServerSideEncryption

Provides information about the default server-side encryption settings for an S3 bucket. For detailed information about these settings, see Setting default server-side encryption behavior for Amazon S3 buckets in the Amazon Simple Storage Service User Guide.

BucketSortCriteria

Specifies criteria for sorting the results of a query for information about S3 buckets.

Cell

Specifies the location of an occurrence of sensitive data in a Microsoft Excel workbook, CSV file, or TSV file.

ClassificationDetails

Provides information about a sensitive data finding, including the classification job that produced the finding.

ClassificationExportConfiguration

Specifies where to store data classification results, and the encryption settings to use when storing results in that location. Currently, you can store classification results only in an S3 bucket.

ClassificationResult

Provides the details of a sensitive data finding, including the types, number of occurrences, and locations of the sensitive data that was detected.

ClassificationResultStatus

Provides information about the status of a sensitive data finding.

CriteriaBlockForJob

Specifies one or more property- and tag-based conditions that define criteria for including or excluding S3 buckets from a classification job.

CriteriaForJob

Specifies a property- or tag-based condition that defines criteria for including or excluding S3 buckets from a classification job.

CriterionAdditionalProperties

Specifies the operator to use in a property-based condition that filters the results of a query for findings. For detailed information and examples of each operator, see Fundamentals of filtering findings in the Amazon Macie User Guide.

CustomDataIdentifierSummary

Provides information about a custom data identifier.

CustomDataIdentifiers

Provides information about custom data identifiers that produced a sensitive data finding, and the number of occurrences of the data that they detected for the finding.

CustomDetection

Provides information about a custom data identifier that produced a sensitive data finding, and the sensitive data that it detected for the finding.

DailySchedule

Specifies that a classification job runs once a day, every day. This is an empty object.

DefaultDetection

Provides information about a type of sensitive data that was detected by managed data identifiers and produced a sensitive data finding.

DomainDetails

Provides information about the domain name of the device that an entity used to perform an action on an affected resource.

FederatedUser

Provides information about an identity that performed an action on an affected resource by using temporary security credentials. The credentials were obtained using the GetFederationToken operation of the Security Token Service (STS) API.

Finding

Provides the details of a finding.

FindingAction

Provides information about an action that occurred for a resource and produced a policy finding.

FindingActor

Provides information about an entity that performed an action that produced a policy finding for a resource.

FindingCriteria

Specifies, as a map, one or more property-based conditions that filter the results of a query for findings.

FindingStatisticsSortCriteria

Specifies criteria for sorting the results of a query that retrieves aggregated statistical data about findings.

FindingsFilterListItem

Provides information about a findings filter.

GroupCount

Provides a group of results for a query that retrieved aggregated statistical data about findings.

IamUser

Provides information about an Identity and Access Management (IAM) user who performed an action on an affected resource.

Invitation

Provides information about an Amazon Macie membership invitation that was received by an account.

IpAddressDetails

Provides information about the IP address of the device that an entity used to perform an action on an affected resource.

IpCity

Provides information about the city that an IP address originated from.

IpCountry

Provides information about the country that an IP address originated from.

IpGeoLocation

Provides geographic coordinates that indicate where a specified IP address originated from.

IpOwner

Provides information about the registered owner of an IP address.

JobDetails

Specifies whether any one-time or recurring classification jobs are configured to analyze data in an S3 bucket, and, if so, the details of the job that ran most recently.

JobScheduleFrequency

Specifies the recurrence pattern for running a classification job.

JobScopeTerm

Specifies a property- or tag-based condition that defines criteria for including or excluding S3 objects from a classification job. A JobScopeTerm object can contain only one simpleScopeTerm object or one tagScopeTerm object.

JobScopingBlock

Specifies one or more property- and tag-based conditions that define criteria for including or excluding S3 objects from a classification job.

JobSummary

Provides information about a classification job, including the current status of the job.

KeyValuePair

Provides information about the tags that are associated with an S3 bucket or object. Each tag consists of a required tag key and an associated tag value.

LastRunErrorStatus

Specifies whether any account- or bucket-level access errors occurred when a classification job ran. For information about using logging data to investigate these errors, see Monitoring sensitive data discovery jobs in the Amazon Macie User Guide.

ListJobsFilterCriteria

Specifies criteria for filtering the results of a request for information about classification jobs.

ListJobsFilterTerm

Specifies a condition that filters the results of a request for information about classification jobs. Each condition consists of a property, an operator, and one or more values.

ListJobsSortCriteria

Specifies criteria for sorting the results of a request for information about classification jobs.

MatchingBucket

Provides statistical data and other information about an S3 bucket that Amazon Macie monitors and analyzes.

MatchingResource

Provides statistical data and other information about an Amazon Web Services resource that Amazon Macie monitors and analyzes.

Member

Provides information about an account that's associated with an Amazon Macie administrator account.

MonthlySchedule

Specifies a monthly recurrence pattern for running a classification job.

ObjectCountByEncryptionType

Provides information about the number of objects that are in an S3 bucket and use certain types of server-side encryption, use client-side encryption, or aren't encrypted.

ObjectLevelStatistics

Provides information about the total storage size (in bytes) or number of objects that Amazon Macie can't analyze in one or more S3 buckets. In a BucketMetadata or MatchingBucket object, this data is for a specific bucket. In a GetBucketStatisticsResponse object, this data is aggregated for all the buckets in the query results. If versioning is enabled for a bucket, total storage size values are based on the size of the latest version of each applicable object in the bucket.

Occurrences

Specifies the location of 1-15 occurrences of sensitive data that was detected by managed data identifiers or a custom data identifier and produced a sensitive data finding.

Page

Specifies the location of an occurrence of sensitive data in an Adobe Portable Document Format file.

PolicyDetails

Provides the details of a policy finding.

Range

Specifies the location of an occurrence of sensitive data in a non-binary text file, such as an HTML, TXT, or XML file.

Record

Specifies the location of an occurrence of sensitive data in an Apache Avro object container, Apache Parquet file, JSON file, or JSON Lines file.

ReplicationDetails

Provides information about settings that define whether one or more objects in an S3 bucket are replicated to S3 buckets for other Amazon Web Services accounts and, if so, which accounts.

ResourcesAffected

Provides information about the resources that a finding applies to.

S3Bucket

Provides information about the S3 bucket that a finding applies to.

S3BucketCriteriaForJob

Specifies property- and tag-based conditions that define criteria for including or excluding S3 buckets from a classification job. Exclude conditions take precedence over include conditions.

S3BucketDefinitionForJob

Specifies an Amazon Web Services account that owns S3 buckets for a classification job to analyze, and one or more specific buckets to analyze for that account.

S3BucketOwner

Provides information about the user who owns an S3 bucket.

S3Destination

Specifies an S3 bucket to store data classification results in, and the encryption settings to use when storing results in that bucket.

S3JobDefinition

Specifies which S3 buckets contain the objects that a classification job analyzes, and the scope of that analysis. The bucket specification can be static (bucketDefinitions) or dynamic (bucketCriteria). If it's static, the job analyzes objects in the same predefined set of buckets each time the job runs. If it's dynamic, the job analyzes objects in any buckets that match the specified criteria each time the job starts to run.

S3Object

Provides information about the S3 object that a finding applies to.

Scoping

Specifies one or more property- and tag-based conditions that define criteria for including or excluding S3 objects from a classification job. Exclude conditions take precedence over include conditions.

SearchResourcesBucketCriteria

Specifies property- and tag-based conditions that define filter criteria for including or excluding S3 buckets from the query results. Exclude conditions take precedence over include conditions.

SearchResourcesCriteria

Specifies a property- or tag-based filter condition for including or excluding Amazon Web Services resources from the query results.

SearchResourcesCriteriaBlock

Specifies property- and tag-based conditions that define filter criteria for including or excluding Amazon Web Services resources from the query results.

SearchResourcesSimpleCriterion

Specifies a property-based filter condition that determines which Amazon Web Services resources are included or excluded from the query results.

SearchResourcesSortCriteria

Specifies criteria for sorting the results of a query for information about Amazon Web Services resources that Amazon Macie monitors and analyzes.

SearchResourcesTagCriterion

Specifies a tag-based filter condition that determines which Amazon Web Services resources are included or excluded from the query results.

SearchResourcesTagCriterionPair

Specifies a tag key, a tag value, or a tag key and value (as a pair) to use in a tag-based filter condition for a query. Tag keys and values are case sensitive. Also, Amazon Macie doesn't support use of partial values or wildcard characters in tag-based filter conditions.

SecurityHubConfiguration

Specifies configuration settings that determine which findings are published to Security Hub automatically. For information about how Macie publishes findings to Security Hub, see Amazon Macie integration with Security Hub in the Amazon Macie User Guide.

SensitiveDataItem

Provides information about the category, types, and occurrences of sensitive data that produced a sensitive data finding.

ServerSideEncryption

Provides information about the server-side encryption settings for an S3 bucket or S3 object.

ServiceLimit

Specifies a current quota for an Amazon Macie account.

SessionContext

Provides information about a session that was created for an entity that performed an action by using temporary security credentials.

SessionContextAttributes

Provides information about the context in which temporary security credentials were issued to an entity.

SessionIssuer

Provides information about the source and type of temporary security credentials that were issued to an entity.

Severity

Provides the numerical and qualitative representations of a finding's severity.

SimpleCriterionForJob

Specifies a property-based condition that determines whether an S3 bucket is included or excluded from a classification job.

SimpleScopeTerm

Specifies a property-based condition that determines whether an S3 object is included or excluded from a classification job.

SortCriteria

Specifies criteria for sorting the results of a request for findings.

Statistics

Provides processing statistics for a classification job.

TagCriterionForJob

Specifies a tag-based condition that determines whether an S3 bucket is included or excluded from a classification job.

TagCriterionPairForJob

Specifies a tag key, a tag value, or a tag key and value (as a pair) to use in a tag-based condition that determines whether an S3 bucket is included or excluded from a classification job. Tag keys and values are case sensitive. Also, Amazon Macie doesn't support use of partial values or wildcard characters in tag-based conditions.

TagScopeTerm

Specifies a tag-based condition that determines whether an S3 object is included or excluded from a classification job.

TagValuePair

Specifies a tag key or tag key and value pair to use in a tag-based condition that determines whether an S3 object is included or excluded from a classification job. Tag keys and values are case sensitive. Also, Amazon Macie doesn't support use of partial values or wildcard characters in tag-based conditions.

UnprocessedAccount

Provides information about an account-related request that hasn't been processed.

UsageByAccount

Provides data for a specific usage metric and the corresponding quota for an Amazon Macie account.

UsageRecord

Provides quota and aggregated usage data for an Amazon Macie account.

UsageStatisticsFilter

Specifies a condition for filtering the results of a query for quota and usage data for one or more Amazon Macie accounts.

UsageStatisticsSortBy

Specifies criteria for sorting the results of a query for Amazon Macie account quotas and usage data.

UsageTotal

Provides aggregated data for an Amazon Macie usage metric. The value for the metric reports estimated usage data for an account for the preceding 30 days or the current calendar month to date, depending on the time period (timeRange) specified in the request.

UserIdentity

Provides information about the type and other characteristics of an entity that performed an action on an affected resource.

UserIdentityRoot

Provides information about an Amazon Web Services account and entity that performed an action on an affected resource. The action was performed using the credentials for your Amazon Web Services account.

UserPausedDetails

Provides information about when a classification job was paused. For a one-time job, this object also specifies when the job will expire and be cancelled if it isn't resumed. For a recurring job, this object also specifies when the paused job run will expire and be cancelled if it isn't resumed. This object is present only if a job's current status (jobStatus) is USER_PAUSED. The information in this object applies only to a job that was paused while it had a status of RUNNING.

WeeklySchedule

Specifies a weekly recurrence pattern for running a classification job.

Enums

AdminStatus

The current status of an account as the delegated Amazon Macie administrator account for an Amazon Web Services organization. Possible values are:

AllowsUnencryptedObjectUploads

NOTE: AllowsUnencryptedObjectUploads::Unknown has been renamed to ::UnknownValue.

Currency

The type of currency that the data for an Amazon Macie usage metric is reported in. Possible values are:

DayOfWeek

EffectivePermission

NOTE: EffectivePermission::Unknown has been renamed to ::UnknownValue.

EncryptionType

The type of server-side encryption that's used to encrypt an S3 object or objects in an S3 bucket. Valid values are:

ErrorCode

The source of an issue or delay. Possible values are:

FindingActionType

The type of action that occurred for the resource and produced the policy finding:

FindingCategory

The category of the finding. Valid values are:

FindingPublishingFrequency

The frequency with which Amazon Macie publishes updates to policy findings for an account. This includes publishing updates to Security Hub and Amazon EventBridge (formerly called Amazon CloudWatch Events). For more information, see Monitoring and processing findings in the Amazon Macie User Guide. Valid values are:

FindingStatisticsSortAttributeName

The grouping to sort the results by. Valid values are:

FindingType

The type of finding. For details about each type, see Types of Amazon Macie findings in the Amazon Macie User Guide. Valid values are:

FindingsFilterAction

The action to perform on findings that meet the filter criteria. To suppress (automatically archive) findings that meet the criteria, set this value to ARCHIVE. Valid values are:

GroupBy

IsDefinedInJob

NOTE: IsDefinedInJob::Unknown has been renamed to ::UnknownValue.

IsMonitoredByJob

NOTE: IsMonitoredByJob::Unknown has been renamed to ::UnknownValue.

JobComparator

The operator to use in a condition. Valid values are:

JobStatus

The status of a classification job. Possible values are:

JobType

The schedule for running a classification job. Valid values are:

LastRunErrorStatusCode

Specifies whether any account- or bucket-level access errors occurred during the run of a one-time classification job or the most recent run of a recurring classification job. Possible values are:

ListJobsFilterKey

The property to use to filter the results. Valid values are:

ListJobsSortAttributeName

The property to sort the results by. Valid values are:

MacieStatus

The status of an Amazon Macie account. Valid values are:

OrderBy

RelationshipStatus

The current status of the relationship between an account and an associated Amazon Macie administrator account (inviter account). Possible values are:

ScopeFilterKey

The property to use in a condition that determines whether an S3 object is included or excluded from a classification job. Valid values are:

SearchResourcesComparator

The operator to use in a condition that filters the results of a query. Valid values are:

SearchResourcesSimpleCriterionKey

The property to use in a condition that filters the query results. Valid values are:

SearchResourcesSortAttributeName

The property to sort the query results by. Valid values are:

SensitiveDataItemCategory

The category of sensitive data that was detected and produced the finding. Possible values are:

SeverityDescription

The qualitative representation of the finding's severity. Possible values are:

SharedAccess

NOTE: SharedAccess::Unknown has been renamed to ::UnknownValue.

SimpleCriterionKeyForJob

The property to use in a condition that determines whether an S3 bucket is included or excluded from a classification job. Valid values are:

StorageClass

The storage class of the S3 object. Possible values are:

TagTarget

The type of object to apply a tag-based condition to. Valid values are:

TimeRange

An inclusive time period that Amazon Macie usage data applies to. Possible values are:

Type

Unit

UsageStatisticsFilterComparator

The operator to use in a condition that filters the results of a query for Amazon Macie account quotas and usage data. Valid values are:

UsageStatisticsFilterKey

The field to use in a condition that filters the results of a query for Amazon Macie account quotas and usage data. Valid values are:

UsageStatisticsSortKey

The field to use to sort the results of a query for Amazon Macie account quotas and usage data. Valid values are:

UsageType

The name of an Amazon Macie usage metric for an account. Possible values are:

UserIdentityType

The type of entity that performed the action on the affected resource. Possible values are: