Struct aws_sdk_macie2::model::bucket_metadata::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for BucketMetadata

Implementations

impl Builder

pub fn account_id(self, input: impl Into<String>) -> Self

The unique identifier for the Amazon Web Services account that owns the bucket.

pub fn set_account_id(self, input: Option<String>) -> Self

The unique identifier for the Amazon Web Services account that owns the bucket.

pub fn allows_unencrypted_object_uploads(
self,
input: AllowsUnencryptedObjectUploads
) -> Self

Specifies whether the bucket policy for the bucket requires server-side encryption of objects when objects are uploaded to the bucket. Possible values are:

FALSE - The bucket policy requires server-side encryption of new objects. PutObject requests must include the x-amz-server-side-encryption header and the value for that header must be AES256 or aws:kms.

TRUE - The bucket doesn't have a bucket policy or it has a bucket policy that doesn't require server-side encryption of new objects. If a bucket policy exists, it doesn't require PutObject requests to include the x-amz-server-side-encryption header and it doesn't require the value for that header to be AES256 or aws:kms.

UNKNOWN - Amazon Macie can't determine whether the bucket policy requires server-side encryption of new objects.

pub fn set_allows_unencrypted_object_uploads(
self,
input: Option<AllowsUnencryptedObjectUploads>
) -> Self

Specifies whether the bucket policy for the bucket requires server-side encryption of objects when objects are uploaded to the bucket. Possible values are:

FALSE - The bucket policy requires server-side encryption of new objects. PutObject requests must include the x-amz-server-side-encryption header and the value for that header must be AES256 or aws:kms.

TRUE - The bucket doesn't have a bucket policy or it has a bucket policy that doesn't require server-side encryption of new objects. If a bucket policy exists, it doesn't require PutObject requests to include the x-amz-server-side-encryption header and it doesn't require the value for that header to be AES256 or aws:kms.

UNKNOWN - Amazon Macie can't determine whether the bucket policy requires server-side encryption of new objects.

pub fn bucket_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the bucket.

pub fn set_bucket_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) of the bucket.

pub fn bucket_created_at(self, input: Instant) -> Self

The date and time, in UTC and extended ISO 8601 format, when the bucket was created.

pub fn set_bucket_created_at(self, input: Option<Instant>) -> Self

The date and time, in UTC and extended ISO 8601 format, when the bucket was created.

pub fn bucket_name(self, input: impl Into<String>) -> Self

The name of the bucket.

pub fn set_bucket_name(self, input: Option<String>) -> Self

The name of the bucket.

pub fn classifiable_object_count(self, input: i64) -> Self

The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

pub fn set_classifiable_object_count(self, input: Option<i64>) -> Self

The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

pub fn classifiable_size_in_bytes(self, input: i64) -> Self

The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

pub fn set_classifiable_size_in_bytes(self, input: Option<i64>) -> Self

The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

pub fn error_code(self, input: BucketMetadataErrorCode) -> Self

Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

pub fn set_error_code(self, input: Option<BucketMetadataErrorCode>) -> Self

Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

pub fn error_message(self, input: impl Into<String>) -> Self

A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

pub fn set_error_message(self, input: Option<String>) -> Self

A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

pub fn job_details(self, input: JobDetails) -> Self

Specifies whether any one-time or recurring classification jobs are configured to analyze data in the bucket, and, if so, the details of the job that ran most recently.

pub fn set_job_details(self, input: Option<JobDetails>) -> Self

Specifies whether any one-time or recurring classification jobs are configured to analyze data in the bucket, and, if so, the details of the job that ran most recently.

pub fn last_updated(self, input: Instant) -> Self

The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the bucket.

pub fn set_last_updated(self, input: Option<Instant>) -> Self

The date and time, in UTC and extended ISO 8601 format, when Amazon Macie most recently retrieved both bucket and object metadata from Amazon S3 for the bucket.

pub fn object_count(self, input: i64) -> Self

The total number of objects in the bucket.

pub fn set_object_count(self, input: Option<i64>) -> Self

The total number of objects in the bucket.

pub fn object_count_by_encryption_type(
self,
input: ObjectCountByEncryptionType
) -> Self

The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

pub fn set_object_count_by_encryption_type(
self,
input: Option<ObjectCountByEncryptionType>
) -> Self

The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

pub fn public_access(self, input: BucketPublicAccess) -> Self

Specifies whether the bucket is publicly accessible due to the combination of permissions settings that apply to the bucket, and provides information about those settings.

pub fn set_public_access(self, input: Option<BucketPublicAccess>) -> Self

Specifies whether the bucket is publicly accessible due to the combination of permissions settings that apply to the bucket, and provides information about those settings.

pub fn region(self, input: impl Into<String>) -> Self

The Amazon Web Services Region that hosts the bucket.

pub fn set_region(self, input: Option<String>) -> Self

The Amazon Web Services Region that hosts the bucket.

pub fn replication_details(self, input: ReplicationDetails) -> Self

Specifies whether the bucket is configured to replicate one or more objects to buckets for other Amazon Web Services accounts and, if so, which accounts.

pub fn set_replication_details(self, input: Option<ReplicationDetails>) -> Self

Specifies whether the bucket is configured to replicate one or more objects to buckets for other Amazon Web Services accounts and, if so, which accounts.

pub fn server_side_encryption(self, input: BucketServerSideEncryption) -> Self

Specifies whether the bucket encrypts new objects by default and, if so, the type of server-side encryption that's used.

pub fn set_server_side_encryption(
self,
input: Option<BucketServerSideEncryption>
) -> Self

Specifies whether the bucket encrypts new objects by default and, if so, the type of server-side encryption that's used.

pub fn shared_access(self, input: SharedAccess) -> Self

Specifies whether the bucket is shared with another Amazon Web Services account. Possible values are:

EXTERNAL - The bucket is shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

INTERNAL - The bucket is shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

NOT_SHARED - The bucket isn't shared with other Amazon Web Services accounts.

UNKNOWN - Amazon Macie wasn't able to evaluate the shared access settings for the bucket.

pub fn set_shared_access(self, input: Option<SharedAccess>) -> Self

Specifies whether the bucket is shared with another Amazon Web Services account. Possible values are:

EXTERNAL - The bucket is shared with an Amazon Web Services account that isn't part of the same Amazon Macie organization.

INTERNAL - The bucket is shared with an Amazon Web Services account that's part of the same Amazon Macie organization.

NOT_SHARED - The bucket isn't shared with other Amazon Web Services accounts.

UNKNOWN - Amazon Macie wasn't able to evaluate the shared access settings for the bucket.

pub fn size_in_bytes(self, input: i64) -> Self

The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

pub fn set_size_in_bytes(self, input: Option<i64>) -> Self

The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

pub fn size_in_bytes_compressed(self, input: i64) -> Self

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

pub fn set_size_in_bytes_compressed(self, input: Option<i64>) -> Self

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

pub fn tags(self, input: impl Into<KeyValuePair>) -> Self

Appends an item to tags.

To override the contents of this collection use set_tags.

An array that specifies the tags (keys and values) that are associated with the bucket.

pub fn set_tags(self, input: Option<Vec<KeyValuePair>>) -> Self

An array that specifies the tags (keys and values) that are associated with the bucket.

pub fn unclassifiable_object_count(self, input: ObjectLevelStatistics) -> Self

The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

pub fn set_unclassifiable_object_count(
self,
input: Option<ObjectLevelStatistics>
) -> Self

The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

pub fn unclassifiable_object_size_in_bytes(
self,
input: ObjectLevelStatistics
) -> Self

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

pub fn set_unclassifiable_object_size_in_bytes(
self,
input: Option<ObjectLevelStatistics>
) -> Self

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

pub fn versioning(self, input: bool) -> Self

Specifies whether versioning is enabled for the bucket.

pub fn set_versioning(self, input: Option<bool>) -> Self

Specifies whether versioning is enabled for the bucket.

pub fn build(self) -> BucketMetadata

Consumes the builder and constructs a BucketMetadata