Struct aws_sdk_macie2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Macie2.

This client allows ergonomic access to a Macie2-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_invitation(&self) -> AcceptInvitation<C, M, R>

Constructs a fluent builder for the AcceptInvitation operation.

See AcceptInvitation for more information about the operation and its arguments.

pub fn batch_get_custom_data_identifiers(
&self
) -> BatchGetCustomDataIdentifiers<C, M, R>

Constructs a fluent builder for the BatchGetCustomDataIdentifiers operation.

See BatchGetCustomDataIdentifiers for more information about the operation and its arguments.

pub fn create_classification_job(&self) -> CreateClassificationJob<C, M, R>

Constructs a fluent builder for the CreateClassificationJob operation.

See CreateClassificationJob for more information about the operation and its arguments.

pub fn create_custom_data_identifier(
&self
) -> CreateCustomDataIdentifier<C, M, R>

Constructs a fluent builder for the CreateCustomDataIdentifier operation.

See CreateCustomDataIdentifier for more information about the operation and its arguments.

pub fn create_findings_filter(&self) -> CreateFindingsFilter<C, M, R>

Constructs a fluent builder for the CreateFindingsFilter operation.

See CreateFindingsFilter for more information about the operation and its arguments.

pub fn create_invitations(&self) -> CreateInvitations<C, M, R>

Constructs a fluent builder for the CreateInvitations operation.

See CreateInvitations for more information about the operation and its arguments.

pub fn create_member(&self) -> CreateMember<C, M, R>

Constructs a fluent builder for the CreateMember operation.

See CreateMember for more information about the operation and its arguments.

pub fn create_sample_findings(&self) -> CreateSampleFindings<C, M, R>

Constructs a fluent builder for the CreateSampleFindings operation.

See CreateSampleFindings for more information about the operation and its arguments.

pub fn decline_invitations(&self) -> DeclineInvitations<C, M, R>

Constructs a fluent builder for the DeclineInvitations operation.

See DeclineInvitations for more information about the operation and its arguments.

pub fn delete_custom_data_identifier(
&self
) -> DeleteCustomDataIdentifier<C, M, R>

Constructs a fluent builder for the DeleteCustomDataIdentifier operation.

See DeleteCustomDataIdentifier for more information about the operation and its arguments.

pub fn delete_findings_filter(&self) -> DeleteFindingsFilter<C, M, R>

Constructs a fluent builder for the DeleteFindingsFilter operation.

See DeleteFindingsFilter for more information about the operation and its arguments.

pub fn delete_invitations(&self) -> DeleteInvitations<C, M, R>

Constructs a fluent builder for the DeleteInvitations operation.

See DeleteInvitations for more information about the operation and its arguments.

pub fn delete_member(&self) -> DeleteMember<C, M, R>

Constructs a fluent builder for the DeleteMember operation.

See DeleteMember for more information about the operation and its arguments.

pub fn describe_buckets(&self) -> DescribeBuckets<C, M, R>

Constructs a fluent builder for the DescribeBuckets operation.

See DescribeBuckets for more information about the operation and its arguments.

pub fn describe_classification_job(&self) -> DescribeClassificationJob<C, M, R>

Constructs a fluent builder for the DescribeClassificationJob operation.

See DescribeClassificationJob for more information about the operation and its arguments.

pub fn describe_organization_configuration(
&self
) -> DescribeOrganizationConfiguration<C, M, R>

Constructs a fluent builder for the DescribeOrganizationConfiguration operation.

See DescribeOrganizationConfiguration for more information about the operation and its arguments.

pub fn disable_macie(&self) -> DisableMacie<C, M, R>

Constructs a fluent builder for the DisableMacie operation.

See DisableMacie for more information about the operation and its arguments.

pub fn disable_organization_admin_account(
&self
) -> DisableOrganizationAdminAccount<C, M, R>

Constructs a fluent builder for the DisableOrganizationAdminAccount operation.

See DisableOrganizationAdminAccount for more information about the operation and its arguments.

pub fn disassociate_from_administrator_account(
&self
) -> DisassociateFromAdministratorAccount<C, M, R>

Constructs a fluent builder for the DisassociateFromAdministratorAccount operation.

See DisassociateFromAdministratorAccount for more information about the operation and its arguments.

pub fn disassociate_from_master_account(
&self
) -> DisassociateFromMasterAccount<C, M, R>

Constructs a fluent builder for the DisassociateFromMasterAccount operation.

See DisassociateFromMasterAccount for more information about the operation and its arguments.

pub fn disassociate_member(&self) -> DisassociateMember<C, M, R>

Constructs a fluent builder for the DisassociateMember operation.

See DisassociateMember for more information about the operation and its arguments.

pub fn enable_macie(&self) -> EnableMacie<C, M, R>

Constructs a fluent builder for the EnableMacie operation.

See EnableMacie for more information about the operation and its arguments.

pub fn enable_organization_admin_account(
&self
) -> EnableOrganizationAdminAccount<C, M, R>

Constructs a fluent builder for the EnableOrganizationAdminAccount operation.

See EnableOrganizationAdminAccount for more information about the operation and its arguments.

pub fn get_administrator_account(&self) -> GetAdministratorAccount<C, M, R>

Constructs a fluent builder for the GetAdministratorAccount operation.

See GetAdministratorAccount for more information about the operation and its arguments.

pub fn get_bucket_statistics(&self) -> GetBucketStatistics<C, M, R>

Constructs a fluent builder for the GetBucketStatistics operation.

See GetBucketStatistics for more information about the operation and its arguments.

pub fn get_classification_export_configuration(
&self
) -> GetClassificationExportConfiguration<C, M, R>

Constructs a fluent builder for the GetClassificationExportConfiguration operation.

See GetClassificationExportConfiguration for more information about the operation and its arguments.

pub fn get_custom_data_identifier(&self) -> GetCustomDataIdentifier<C, M, R>

Constructs a fluent builder for the GetCustomDataIdentifier operation.

See GetCustomDataIdentifier for more information about the operation and its arguments.

pub fn get_findings(&self) -> GetFindings<C, M, R>

Constructs a fluent builder for the GetFindings operation.

See GetFindings for more information about the operation and its arguments.

pub fn get_findings_filter(&self) -> GetFindingsFilter<C, M, R>

Constructs a fluent builder for the GetFindingsFilter operation.

See GetFindingsFilter for more information about the operation and its arguments.

pub fn get_findings_publication_configuration(
&self
) -> GetFindingsPublicationConfiguration<C, M, R>

Constructs a fluent builder for the GetFindingsPublicationConfiguration operation.

See GetFindingsPublicationConfiguration for more information about the operation and its arguments.

pub fn get_finding_statistics(&self) -> GetFindingStatistics<C, M, R>

Constructs a fluent builder for the GetFindingStatistics operation.

See GetFindingStatistics for more information about the operation and its arguments.

pub fn get_invitations_count(&self) -> GetInvitationsCount<C, M, R>

Constructs a fluent builder for the GetInvitationsCount operation.

See GetInvitationsCount for more information about the operation and its arguments.

pub fn get_macie_session(&self) -> GetMacieSession<C, M, R>

Constructs a fluent builder for the GetMacieSession operation.

See GetMacieSession for more information about the operation and its arguments.

pub fn get_master_account(&self) -> GetMasterAccount<C, M, R>

Constructs a fluent builder for the GetMasterAccount operation.

See GetMasterAccount for more information about the operation and its arguments.

pub fn get_member(&self) -> GetMember<C, M, R>

Constructs a fluent builder for the GetMember operation.

See GetMember for more information about the operation and its arguments.

pub fn get_usage_statistics(&self) -> GetUsageStatistics<C, M, R>

Constructs a fluent builder for the GetUsageStatistics operation.

See GetUsageStatistics for more information about the operation and its arguments.

pub fn get_usage_totals(&self) -> GetUsageTotals<C, M, R>

Constructs a fluent builder for the GetUsageTotals operation.

See GetUsageTotals for more information about the operation and its arguments.

pub fn list_classification_jobs(&self) -> ListClassificationJobs<C, M, R>

Constructs a fluent builder for the ListClassificationJobs operation.

See ListClassificationJobs for more information about the operation and its arguments.

pub fn list_custom_data_identifiers(&self) -> ListCustomDataIdentifiers<C, M, R>

Constructs a fluent builder for the ListCustomDataIdentifiers operation.

See ListCustomDataIdentifiers for more information about the operation and its arguments.

pub fn list_findings(&self) -> ListFindings<C, M, R>

Constructs a fluent builder for the ListFindings operation.

See ListFindings for more information about the operation and its arguments.

pub fn list_findings_filters(&self) -> ListFindingsFilters<C, M, R>

Constructs a fluent builder for the ListFindingsFilters operation.

See ListFindingsFilters for more information about the operation and its arguments.

pub fn list_invitations(&self) -> ListInvitations<C, M, R>

Constructs a fluent builder for the ListInvitations operation.

See ListInvitations for more information about the operation and its arguments.

pub fn list_managed_data_identifiers(
&self
) -> ListManagedDataIdentifiers<C, M, R>

Constructs a fluent builder for the ListManagedDataIdentifiers operation.

See ListManagedDataIdentifiers for more information about the operation and its arguments.

pub fn list_members(&self) -> ListMembers<C, M, R>

Constructs a fluent builder for the ListMembers operation.

See ListMembers for more information about the operation and its arguments.

pub fn list_organization_admin_accounts(
&self
) -> ListOrganizationAdminAccounts<C, M, R>

Constructs a fluent builder for the ListOrganizationAdminAccounts operation.

See ListOrganizationAdminAccounts for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn put_classification_export_configuration(
&self
) -> PutClassificationExportConfiguration<C, M, R>

Constructs a fluent builder for the PutClassificationExportConfiguration operation.

See PutClassificationExportConfiguration for more information about the operation and its arguments.

pub fn put_findings_publication_configuration(
&self
) -> PutFindingsPublicationConfiguration<C, M, R>

Constructs a fluent builder for the PutFindingsPublicationConfiguration operation.

See PutFindingsPublicationConfiguration for more information about the operation and its arguments.

pub fn search_resources(&self) -> SearchResources<C, M, R>

Constructs a fluent builder for the SearchResources operation.

See SearchResources for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn test_custom_data_identifier(&self) -> TestCustomDataIdentifier<C, M, R>

Constructs a fluent builder for the TestCustomDataIdentifier operation.

See TestCustomDataIdentifier for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_classification_job(&self) -> UpdateClassificationJob<C, M, R>

Constructs a fluent builder for the UpdateClassificationJob operation.

See UpdateClassificationJob for more information about the operation and its arguments.

pub fn update_findings_filter(&self) -> UpdateFindingsFilter<C, M, R>

Constructs a fluent builder for the UpdateFindingsFilter operation.

See UpdateFindingsFilter for more information about the operation and its arguments.

pub fn update_macie_session(&self) -> UpdateMacieSession<C, M, R>

Constructs a fluent builder for the UpdateMacieSession operation.

See UpdateMacieSession for more information about the operation and its arguments.

pub fn update_member_session(&self) -> UpdateMemberSession<C, M, R>

Constructs a fluent builder for the UpdateMemberSession operation.

See UpdateMemberSession for more information about the operation and its arguments.

pub fn update_organization_configuration(
&self
) -> UpdateOrganizationConfiguration<C, M, R>

Constructs a fluent builder for the UpdateOrganizationConfiguration operation.

See UpdateOrganizationConfiguration for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.